fixup! feat(webdav): initial implementation for a WebDAV simple auth provider
